services/support_services/azure_ad_sso_service/requirements.txt
# Azure AD SSO Service Dependencies

# HTTP Client
httpx>=0.24.0

# JWT Token Handling
pyjwt[crypto]>=2.8.0
cryptography>=41.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.28.0

# Data Validation
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Async Support
asyncio>=3.4.3

# Date/Time
python-dateutil>=2.8.2

# Logging
python-json-logger>=2.0.7
